refactor: ApiErrorCode 패키지 변경
