+++ src/main/java/ktb3/full/week04/exception/ApiErrorCode.java
@@ -1,4 +1,4 @@
-package ktb3.full.week04.dto.response;
+package ktb3.full.week04.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
