+++ src/main/java/ktb3/full/week04/exception/InvalidInputException.java
@@ -1,6 +1,5 @@
 package ktb3.full.week04.exception;
 
-import ktb3.full.week04.dto.response.ApiErrorCode;
 import ktb3.full.week04.exception.base.BadRequestException;
 
 public class InvalidInputException extends BadRequestException {
