+++ src/main/java/ktb3/full/week04/exception/LoginRequiredException.java
@@ -1,6 +1,5 @@
 package ktb3.full.week04.exception;
 
-import ktb3.full.week04.dto.response.ApiErrorCode;
 import ktb3.full.week04.exception.base.UnauthorizedException;
 
 public class LoginRequiredException extends UnauthorizedException {
