+++ src/main/java/ktb3/full/week04/exception/base/CustomException.java
@@ -1,6 +1,6 @@
 package ktb3.full.week04.exception.base;
 
-import ktb3.full.week04.dto.response.ApiErrorCode;
+import ktb3.full.week04.exception.ApiErrorCode;
 import org.springframework.http.HttpStatus;
 
 public abstract class CustomException extends RuntimeException {
