+++ src/main/java/ktb3/full/week04/exception/handler/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
 package ktb3.full.week04.exception.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
-import ktb3.full.week04.dto.response.ApiErrorCode;
+import ktb3.full.week04.exception.ApiErrorCode;
 import ktb3.full.week04.dto.response.ApiErrorResponse;
 import ktb3.full.week04.exception.base.CustomException;
 import lombok.extern.slf4j.Slf4j;
